implemented a check in every codegen to use option's default value if it isn't submitted by user, thus getOptions works as source of truth for defaults
