+++ codegens/js-fetch/lib/index.js
@@ -179,6 +179,11 @@ function convert (request, options, callback) {
     if (!_.isFunction(callback)) {
         throw new Error('JS-Fetch Converter callback is not a valid function');
     }
+    getOptions().forEach((option) => {
+        if (_.isUndefined(options[option.id])) {
+            options[option.id] = option.default;
+        }
+    });
 
     var indent = options.indentType === 'tab' ? '\t' : ' ',
         trim = options.trimRequestBody,
