+++ codegens/js-jquery/lib/js-jquery.js
@@ -1,6 +1,7 @@
 var _ = require('./lodash'),
     parseBody = require('./util/parseBody'),
-    sanitize = require('./util/sanitize').sanitize;
+    sanitize = require('./util/sanitize').sanitize,
+    self;
 
 /**
      * Used to parse the request headers
@@ -47,7 +48,7 @@ function createForm (request, trimRequestBody) {
     return form;
 }
 
-module.exports = {
+self = module.exports = {
     /**
      * Used to return options which are specific to a particular plugin
      *
@@ -110,11 +111,14 @@ module.exports = {
         else if (!_.isFunction(callback)) {
             throw new Error('js-jQuery~convert: Callback is not a function');
         }
-
+        self.getOptions().forEach((option) => {
+            if (_.isUndefined(options[option.id])) {
+                options[option.id] = option.default;
+            }
+        });
         indentType = (options.indentType === 'tab') ? '\t' : ' ';
 
-        indent = indentType.repeat(options.indentCount || (options.indentType === 'tab' ? 1 : 4));
-        options.requestTimeout = options.requestTimeout || 0;
+        indent = indentType.repeat(options.indentCount);
 
         if (request.body && request.body.mode === 'formdata') {
             jQueryCode = createForm(request.toJSON(), options.trimRequestBody);
