+++ codegens/nodejs-request/lib/request.js
@@ -5,11 +5,11 @@ var _ = require('./lodash'),
 
 /**
  * retuns snippet of nodejs(request) by parsing data from Postman-SDK request object
- * 
+ *
  * @param {Object} request - Postman SDK request object
  * @param {String} indentString - indentation required for code snippet
  * @param {Object} options
- * @returns {String} - nodejs(request) code snippet for given request object 
+ * @returns {String} - nodejs(request) code snippet for given request object
  */
 function makeSnippet (request, indentString, options) {
     var snippet = 'var request = require(\'request\');\n',
@@ -18,15 +18,15 @@ function makeSnippet (request, indentString, options) {
     snippet += 'var fs = require(\'fs\')\n';
     snippet += 'var options = {\n';
 
-    /** 
+    /**
      * creating string to represent options object using optionArray.join()
      * example:
      *  options: {
      *      method: 'GET',
      *      url: 'www.google.com',
-     *      timeout: 1000   
+     *      timeout: 1000
      *  }
-     */ 
+     */
     optionsArray.push(indentString + `'method': '${request.method}'`);
     optionsArray.push(indentString + `'url': '${sanitize(request.url.toString())}'`);
 
@@ -102,8 +102,8 @@ function getOptions () {
 
 /**
  * Converts Postman sdk request object to nodejs request code snippet
- * 
- * @param {Object} request - postman-SDK request object 
+ *
+ * @param {Object} request - postman-SDK request object
  * @param {Object} options
  * @param {String} options.indentType - type for indentation eg: space, tab
  * @param {String} options.indentCount - number of spaces or tabs for indentation.
@@ -116,6 +116,11 @@ function convert (request, options, callback) {
     if (!_.isFunction(callback)) {
         throw new Error('NodeJS-Request-Converter: callback is not valid function');
     }
+    getOptions().forEach((option) => {
+        if (_.isUndefined(options[option.id])) {
+            options[option.id] = option.default;
+        }
+    });
 
     //  String representing value of indentation required
     var indentString;
