+++ codegens/nodejs-unirest/lib/unirest.js
@@ -4,11 +4,11 @@ var _ = require('./lodash'),
 
 /**
  * retuns snippet of nodejs(unirest) by parsing data from Postman-SDK request object
- * 
+ *
  * @param {Object} request - Postman SDK request object
  * @param {String} indentString - indentation required for code snippet
  * @param {Object} options
- * @returns {String} - nodejs(unirest) code snippet for given request object 
+ * @returns {String} - nodejs(unirest) code snippet for given request object
  */
 function makeSnippet (request, indentString, options) {
     var snippet = 'var unirest = require(\'unirest\');\n';
@@ -83,8 +83,8 @@ function getOptions () {
 
 /**
  * Converts Postman sdk request object to nodejs(unirest) code snippet
- * 
- * @param {Object} request - postman-SDK request object 
+ *
+ * @param {Object} request - postman-SDK request object
  * @param {Object} options
  * @param {String} options.indentType - type for indentation eg: space, tab
  * @param {String} options.indentCount - number of spaces or tabs for indentation.
@@ -97,6 +97,11 @@ function convert (request, options, callback) {
     if (!_.isFunction(callback)) {
         throw new Error('Nodejs-Unirest-Converter: callback is not valid function');
     }
+    getOptions().forEach((option) => {
+        if (_.isUndefined(options[option.id])) {
+            options[option.id] = option.default;
+        }
+    });
 
     //  String representing value of indentation required
     var indentString;
