+++ codegens/swift/lib/swift.js
@@ -1,5 +1,6 @@
 var _ = require('./lodash'),
-    sanitize = require('./util').sanitize;
+    sanitize = require('./util').sanitize,
+    self;
 
 /**
  * Parses Raw data from request to fetch syntax
@@ -140,7 +141,7 @@ function parseHeaders (headers, mode) {
     return headerSnippet;
 }
 
-module.exports = {
+self = module.exports = {
     /**
      * Used in order to get additional options for generation of Swift code snippet
      *
@@ -215,14 +216,18 @@ module.exports = {
         else if (!_.isFunction(callback)) {
             throw new Error('Swift-Converter: callback is not valid function');
         }
-
+        self.getOptions().forEach((option) => {
+            if (_.isUndefined(options[option.id])) {
+                options[option.id] = option.default;
+            }
+        });
         var codeSnippet, indent, trim, timeout, finalUrl, // followRedirect,
             bodySnippet = '',
             headerSnippet = '',
             requestBody = (request.body ? request.body.toJSON() : {});
 
         indent = options.indentType === 'tab' ? '\t' : ' ';
-        indent = indent.repeat(options.indentCount || (options.indentType === 'tab' ? 1 : 4));
+        indent = indent.repeat(options.indentCount);
         timeout = options.requestTimeout;
         // followRedirect = options.followRedirect;
         trim = options.trimRequestBody;
